+++ login-workflow/jestSetupFile.js
@@ -1,6 +1,6 @@
 import mockAsyncStorage from '@react-native-async-storage/async-storage/jest/async-storage-mock';
-jest.mock('react-native/Libraries/Animated/src/NativeAnimatedHelper');
 jest.mock('@react-native-async-storage/async-storage', () => mockAsyncStorage);
+jest.mock('react-native/Libraries/Animated/NativeAnimatedHelper');
 jest.mock('react-i18next', () => ({
     useTranslation: () => ({ t: (key) => key }),
 }));